feat: add built-in Classifier blueprint for object instantiation
Add Classifier blueprint that creates instances from module paths and
kwargs. Register it for object type as fallback for non-Pydantic
classes. Move from test file to dedicated blueprints module and
auto-import to ensure registration.
