+++ cyantic/__init__.py
@@ -2,6 +2,9 @@ from cyantic.core import blueprint, Blueprint, CyanticModel
 from cyantic.context import ValidationContext
 from cyantic.hooks import hook, HookRegistry
 
+# Import blueprints module to register built-in blueprints
+from cyantic import blueprints  # noqa: F401
+
 __all__ = [
     "blueprint",
     "Blueprint",
